requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
pytest>=8.0
PyYAML>=6.0
